move DeviceCapabilitiesExtension into its own file
